# Reduce repeat allocations of ZeekString/StringVal objects