--- a/src/plugin/Component.cc
+++ b/src/plugin/Component.cc
@@ -16,6 +16,7 @@ Component::Component(component::Type arg_type, const std::string& arg_name,
 	  tag_subtype(tag_subtype)
 	{
 	canon_name = util::canonify_name(name);
+	canon_name_val = make_intrusive<StringVal>(canon_name);
 	}
 
 void Component::Describe(ODesc* d) const