Fixes #1797

This PR fixes 3 issues that cause large numbers of repeat allocations of the same strings into ZeekString/StringVal objects:
- Connection history
- Analyzer canonical names
- Hash analyzer kinds

From testing with a pcap with 1 million HTTP connections, we get the following allocations for connection history strings (amongst others):
```
1830308 A Allocating a string of len 6: ShADad
1252316 A Allocating a string of len 4: ShAD
```

With the patch in https://github.com/zeek/zeek/commit/1f8d40665846c262e0fcd197a4b9811618e55c0e, these are reduced to:
```
96332 A Allocating a string of len 6: ShADad
96332 A Allocating a string of len 4: ShAD
```

From testing with a pcap with a very large number of SSL connections, we get the following for various analyzers:

```
  43369 A Allocating a string of len 3: SSL
  23460 A Allocating a string of len 4: sha1
  23457 A Allocating a string of len 6: sha256
  23457 A Allocating a string of len 3: md5
  23456 A Allocating a string of len 6: SHA256
  23456 A Allocating a string of len 4: X509
  23456 A Allocating a string of len 4: SHA1
  23456 A Allocating a string of len 3: MD5
```

With the patches in https://github.com/zeek/zeek/commit/a680c2faf004601be51db2611b379e8f25a12b7a and https://github.com/zeek/zeek/commit/b850d1dc515898df211234ba1311c01bf13245ab, these are reduced to:

```
  33415 A Allocating a string of len 3: SSL
      5 A Allocating a string of len 4: sha1
      1 A Allocating a string of len 4: SHA1
      2 A Allocating a string of len 3: md5
      1 A Allocating a string of len 3: MD5
      2 A Allocating a string of len 6: sha256
      1 A Allocating a string of len 6: SHA256
      1 A Allocating a string of len 4: X509
```

https://github.com/zeek/zeek/commit/a680c2faf004601be51db2611b379e8f25a12b7a causes a smattering of extra allocations that don't show up in the unpatched version of the tests, but it's basically 1 or 2 allocations per analyzer/component vs an extra 23000+.

Surprisingly the time savings isn’t as huge as expected, but it does exist. Averaged over 5 reads of `http_many.pcap`, runtime changed from 44.87s to 43.91s (~2%). Averaged over 10 reads of `ssl_many.pcap`, runtime changed from 8.25s to 8.12s (~1.5%). I’ll have longer benchmarks shortly.